requirements.txt
mysql-connector==2.1.7
requests
pytz
python-multipart
starlette
xmltodict